Improve keyring handling
The script will now use the default keyring per default. If no default
keyring is set, it tries to create one. If this behavior is not desired,
the script now includes two parameters ({use_default}, {keyring_name})
that allow to use a different keyring.

With the new behavior, the used keyring is also unlocked automatically
when starting the script.